# Shortcut method to repack file segments using copy_file_range