This should speed up repacking file segments.

I may need help testing it. I'll now try to download a dataset with large files to test it locally, though it would help to test it on throwaway datasets on ARPA machines