@spanezz 2025-01-23
I created a tool that inverts the order of data in a segment, therefore rewriting it completely. It's a worst-case scenario for the new code, which is faster if large chunks of the segment remain the same.

In my laptop, on a 389M segment from lama5, the time it takes to rewrite it goes from 1.60s to 0.9s. It's almost half, and I'd say it's a saving worth looking into.

`arki-check --accurate` in my system seems happy after several rewrites.

I tried a repack after removing a product from all online segments of a lama5 copy, and it went from 5.84s to 2.81s, which is a bit less than half the time.